Consolidate CosmosBiRelay constructs into one file
