--- crates/cosmos/cosmos-relayer/src/impls/birelay/mod.rs
@@ -1,2 +1,0 @@
-pub mod components;
-pub mod types;
